Group Docker builds by buildpack (#5)
This minimizes the number of images that we need to build. For example, if you have 10 tasks that all use the same Dockerfile, then that image will only be built once (but be tagged 10 times).

This PR also adds a `parallel` flag that can be disabled to make it easier to debug output (without docker builds getting interspersed with each other).

It also patches a bug with the Docker buildpack.