Tag with :latest on default branch (#9)
We will now publish to the `:latest` tag on your default branch, which is the tag that builders pull from by default now.

For customers whose default branch is `main` or `master`, they will not have to do anything. For all new customers, the default workflow will include a `default-branch: $default-branch` input which will pass in the correct branch.

Will release as v0.1.7.