Add node builder (#13)
This builder takes in an entrypoint and finds a working directory based
on the nearest directory with package.json.

From there, if package-lock.json is found, it'll use npm install. Or
else, it'll use yarn.

It'll use typescript's tsc to compile to a build directory, and then
sets a Docker entrypoint running the entrypoint script out of the build
directory.